requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
blinker==1.6.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
flask-restplus==0.13.0
flask-restx==1.1.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
pycparser==2.20
PyJWT==2.7.0
PyMySQL==1.0.2
pyrsistent==0.18.0
pytz==2021.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.3.3
zipp==3.5.0
pycryptodomex==3.18.0
gspread==5.10.0
python-memcached==1.59
pymysql-pooling==1.0.6
APScheduler==3.10.4
backports.zoneinfo==0.2.1
tzlocal==5.2
firebase_admin==6.3.0
